requirements.txt
discord.py>=2.0.0
aiohttp>=3.8.0
aiofiles>=0.8.0
python-dotenv>=0.19.0
cryptography>=36.0.0